Logging cleanup
The Config section is now independent of the File and debug mode management.
Errors are no longer logged to console.
Native errors and their traces are logged to file
A mechanism to save UI logs is pending.

Some privacy violations might exist in the case of logging file locations: this might expose a subset of the users directory structure.